Disable lots of unused protocols and rerun androidconfigure.
These less used protocol that we don't need will often contain bugs we
don't want. Stop compiling them.

Rearranged the ./configure args into an array to allow documenation of
the comments and udpated the path to the lib/ directory to make
./configure happy. Updated the local-configure.patch to reflect changes
in the curl_config.h file. Several types are now unused so the patch is
different. We still remove those from the curl_configure.h file even if
they are just "#undef" comments to conflict (and fail androidconfigure
command) if somebody ever needs them in the future. These values are
defined in the Android.mk instead.

The ./androidconfigure command now shows:
  Protocols:        FILE FTP FTPS HTTP HTTPS

Bug: 69830574
Test: mmma external/libcurl; Inspected curl_config.h changes.
Change-Id: If9f4ce09ef54e9ba82fd725ab11224a92d794535